telekinesis fixes and sound
