+++ Content.Shared/Weapons/Misc/COSharedTelekinesisSystem.cs
@@ -3,12 +3,14 @@ using Content.Shared.Hands.EntitySystems;
 using Content.Shared.Popups;
 using Content.Shared.Tag;
 using Content.Shared.Toggleable;
+using Robust.Shared.Network;
 using Robust.Shared.Prototypes;
 
 namespace Content.Shared.Weapons.Misc;
 
 public abstract class COSharedTelekinesisSystem : EntitySystem
 {
+    [Dependency] private readonly INetManager _netManager = default!;
     [Dependency] protected readonly SharedTransformSystem TransformSystem = default!;
     [Dependency] protected readonly SharedHandsSystem _hands = default!;
     [Dependency] protected readonly TagSystem _tagSystem = default!;
@@ -25,6 +27,27 @@ public abstract class COSharedTelekinesisSystem : EntitySystem
     {
         base.Initialize();
         SubscribeLocalEvent<COTelekinesisComponent, MapInitEvent>(OnMapInit);
+        SubscribeLocalEvent<COTelekinesisComponent, ToggleActionEvent>(TelekinesisAction);
+    }
+
+    private void TelekinesisAction(Entity<COTelekinesisComponent> ent, ref ToggleActionEvent args)
+    {
+        if (args.Handled)
+            return;
+
+        TryTelekinesisTether((ent.Owner, ent.Comp), args.Performer);
+        args.Handled = true;
+    }
+
+    private bool TryTelekinesisTether(Entity<COTelekinesisComponent?> ent, EntityUid user)
+    {
+        if (!Resolve(ent, ref ent.Comp, false))
+            return false;
+
+        bool enabled = TryTether((ent.Owner, ent.Comp), user);
+
+        _actionsSystem.SetToggled(ent.Comp.Action, enabled);
+        return enabled;
     }
 
     private void OnMapInit(Entity<COTelekinesisComponent> ent, ref MapInitEvent args)
@@ -42,11 +65,10 @@ public abstract class COSharedTelekinesisSystem : EntitySystem
         if (telekinesis is not null)
         {
             StopTether(telekinesis);
-
             return false;
         }
 
-        StartTether(user);
+        StartTether(ent, user);
         return true;
     }
 
@@ -69,7 +91,7 @@ public abstract class COSharedTelekinesisSystem : EntitySystem
         return tetherPower;
     }
 
-    protected virtual void StartTether(EntityUid user)
+    protected virtual void StartTether(Entity<COTelekinesisComponent?> ent, EntityUid user)
     { }
 
     protected virtual void StopTether(EntityUid? telekinesis)
