+++ Content.Shared/Weapons/Misc/COTelekinesisComponent.cs
@@ -1,3 +1,4 @@
+using Robust.Shared.Audio;
 using Robust.Shared.GameStates;
 using Robust.Shared.Prototypes;
 
@@ -6,8 +7,10 @@ namespace Content.Shared.Weapons.Misc;
 [RegisterComponent, NetworkedComponent, AutoGenerateComponentState, Access(typeof(COSharedTelekinesisSystem))]
 public sealed partial class COTelekinesisComponent : Component
 {
-    [ViewVariables, DataField, AutoNetworkedField]
+    [DataField]
+    public SoundSpecifier ActivateSound = new SoundPathSpecifier("/Audio/Magic/staff_animation.ogg");
 
+    [ViewVariables, DataField, AutoNetworkedField]
     public EntProtoId ActionProto = "COActionTelekinesis";
 
     [DataField, AutoNetworkedField]
